backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
PyPDF2==3.0.1
